Move cmd functions to hooks
