+++ packages/teleport/src/Apps/AddApp/useAddApp.ts
@@ -14,9 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
+import { formatDistanceStrict } from 'date-fns';
 import useAttempt from 'shared/hooks/useAttemptNext';
-import { createAppBashCommand } from 'teleport/services/commands';
+import { BashCommand, NodeToken } from 'teleport/services/nodes';
+import cfg from 'teleport/config';
 import TeleportContext from 'teleport/teleportContext';
 
 export default function useAddApp(ctx: TeleportContext) {
@@ -53,6 +55,45 @@ export default function useAddApp(ctx: TeleportContext) {
     isAuthTypeLocal,
     isEnterprise,
     token,
+  };
+}
+
+export function createAppBashCommand(
+  token: NodeToken,
+  appName = '',
+  appUri = ''
+): BashCommand {
+  const expires = formatDistanceStrict(new Date(), token.expiry);
+
+  // encode uri so it can be passed around as URL query parameter
+  const encoded = encodeURIComponent(appUri)
+    // encode single quotes so they do not break the curl parameters
+    .replace(/'/g, '%27');
+
+  const bashUrl =
+    cfg.baseUrl +
+    cfg.api.appNodeScriptPath
+      .replace(':token', token.id)
+      .replace(':name', appName)
+      .replace(':uri', encoded);
+
+  const text = `sudo bash -c "$(curl -fsSL '${bashUrl}')"`;
+
+  return {
+    text,
+    expires,
+  };
+}
+
+export function createNodeBashCommand(node: NodeToken): BashCommand {
+  const { expiry, id } = node;
+
+  const expires = formatDistanceStrict(new Date(), new Date(expiry));
+  const text = `sudo bash -c "$(curl -fsSL ${cfg.getNodeScriptUrl(id)})"`;
+
+  return {
+    text,
+    expires,
   };
 }
