+++ packages/teleport/src/Nodes/AddNode/useAddNode.ts
@@ -15,9 +15,11 @@ limitations under the License.
 */
 
 import { useState, useEffect } from 'react';
+import { formatDistanceStrict } from 'date-fns';
 import useAttempt from 'shared/hooks/useAttemptNext';
 import TeleportContext from 'teleport/teleportContext';
-import { createNodeBashCommand } from 'teleport/services/commands';
+import { BashCommand, NodeToken } from 'teleport/services/nodes';
+import cfg from 'teleport/config';
 
 export default function useAddNode(ctx: TeleportContext) {
   const { attempt, run } = useAttempt('processing');
@@ -57,6 +59,18 @@ export default function useAddNode(ctx: TeleportContext) {
     user,
     isAuthTypeLocal,
     token,
+  };
+}
+
+export function createNodeBashCommand(node: NodeToken): BashCommand {
+  const { expiry, id } = node;
+
+  const expires = formatDistanceStrict(new Date(), new Date(expiry));
+  const text = `sudo bash -c "$(curl -fsSL ${cfg.getNodeScriptUrl(id)})"`;
+
+  return {
+    text,
+    expires,
   };
 }
